custom_vision/requirements.txt
azure-cognitiveservices-vision-customvisin==3.1.1
msrest==0.7.
pandas==2.2.3
matplotlib==3.10.0
numpy==2.2.1
Pillow==11.0.0
python-dotenv==1.0.1